cloud_run/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.1.3
openpyxl==3.1.2
duckdb==0.9.2
python-dotenv==1.0.0
google-cloud-storage==2.10.0